+++ src/utils.rs
@@ -234,15 +234,16 @@ mod tests {
     fn test_extract_expressions() {
         let multiple = r#"echo "OSSL_PATH=${{ github.workspace }}/osslcache/${{ matrix.PYTHON.OPENSSL.TYPE }}-${{ matrix.PYTHON.OPENSSL.VERSION }}-${OPENSSL_HASH}" >> $GITHUB_ENV"#;
 
-        for (raw, expected) in &[(
-            multiple,
-            [
-                "${{ github.workspace }}",
-                "${{ matrix.PYTHON.OPENSSL.TYPE }}",
-                "${{ matrix.PYTHON.OPENSSL.VERSION }}",
-            ]
-            .as_slice(),
-        )] {
+        {
+            let (raw, expected) = &(
+                multiple,
+                [
+                    "${{ github.workspace }}",
+                    "${{ matrix.PYTHON.OPENSSL.TYPE }}",
+                    "${{ matrix.PYTHON.OPENSSL.VERSION }}",
+                ]
+                .as_slice(),
+            );
             let exprs = extract_expressions(raw)
                 .into_iter()
                 .map(|(e, _)| e.as_curly().to_string())